\name{is.S}
\alias{is.S}
\alias{is.Splus}
\alias{is.Svanilla}
\alias{is.Splus.pre3.3}

\title{S/R Version Information}
\description{Determine the S/R version.}

\usage{
    is.S()
    is.Splus()
    is.Svanilla()
    is.Splus.pre3.3()
}

\arguments{None}
\value{TRUE or FALSE}
\details{
    If the S/R version is the one in question then TRUE is
    returned, otherwise FALSE. 
}
\seealso{
    \code{\link{is.R}}
}
\examples{is.R() }
\keyword{programming}
\keyword{utilities}

