\name{SPAWN}
\alias{.SPAWN}
\title{Global variable}
\description{Global variable (can be used to control "For loops" in Splus).}
\usage{
	.SPAWN <- FALSE}
\details{
	Functions coded to use Splus "For loops" can check this global
	variable to see if they should use them or not. 
}
\keyword{syskern}
\keyword{OS}
\keyword{S/R}
