\name{system.call}
\alias{system.call}
\alias{sleep}
\alias{mail}
\alias{file.copy}
\alias{file.date.info}

\title{Operating system requests}
\description{Various requests to the operating system.}

\usage{
	system.call()
	sleep()
	mail()
	file.copy()
	file.date.info()
}

\arguments{...}
\value{misc.}
\details{
	Various requests to the operating system. 
}
\seealso{
	\code{\link{system.info}}
}
\examples{system.call("ls") }
\keyword{syskern}
\keyword{OS}
\keyword{S/R}

