\name{system.info}
\alias{system.info}

\title{Get information about the system}
\description{Get information about the system.}

\usage{
	system.info()}

\arguments{None}
\value{A list.}
\details{
	Return information about the system. 
}
\seealso{
	\code{\link{system.call}}
	\code{\link{is.S}}
}
\examples{system.info() }
\keyword{syskern}
\keyword{OS}
\keyword{S/R}

