\name{whoami}
\alias{whoami}
\alias{local.host.netname}
\alias{OStype}
\alias{OSversion}

\title{Operating System Information}
\description{Get information from the operating system.}

\usage{
	whoami()
	getwd()
	local.host.netname()
	OStype()
	OSversion()
}

\arguments{None}
\value{A character string.}
\details{
	Return requested information from the operating system. 
}
\seealso{
	\code{\link{system.info}}
	\code{\link{system.call}}
}
\examples{whoami() }
\keyword{syskern}
\keyword{OS}
\keyword{S/R}

