\name{global.assign}
\alias{global.assign}
\alias{exists.graphics.device}
\alias{open.graphics.device}
\alias{dev.ask}
\alias{list.add<-}
\alias{date.parsed}
\alias{synchronize}

\title{Utilities to handle small S/R differences}
\description{Utilities to handle small S/R differences.}

\usage{
	global.assign()
	exists.graphics.device()
	open.graphics.device()
	global.assign()
	dev.ask()
	date.parsed()
	synchronize()
}

\arguments{misc.}
\value{misc.}
\details{
	Utilities to handle small S/R differences. 
}
\seealso{
	\code{\link{set.RNG}}
	\code{\link{is.S}}
}
\examples{date.parsed() }
\keyword{syskern}
\keyword{OS}
\keyword{S/R}

